Moderation page and admin redesign
This commit includes quite a lot of changes and has taken some time for me to put together. It includes changes to the API backend in addition to frontend work. Responsiveness might not work the best but it should be sufficient for the two pages. The two /admin and /mods routes are accessible through the account dropdown in the navbar when signed in. The staff cards aren't currently editable but will display the current user's card (or at least it should, if they don't exist in the file it might break the page). The current and future staff card data will need to be put in the database for easy management of it. I can do this locally but it will also need to be done for production.