Fix issue preventing users signing in
This is a bug likely caused by Discord sending the requested scopes string back in the wrong order or potentially the server incorrectly disposing of an old cookie. Either way this should fix it by instead of directly comparing the strings to split the received scopes into an array and check if it includes what we need.